Secondary Resource Crafting Fixes and Refactoring
Secondary resource using recipes were refactored so that
you state what the secondary resource is in one of the recipe
text files at the recipe's entry by appending the secondary
resource to the entry in three possible ways:

	a. If the secondary resource is a material stack,
	   you specify the name of the material and the
	   amount consumed from the stack.

	b. If the secondary resource is a non-material stack,
	   you specify the type of the stack, a name you want
	   to be displayed for it and the amount consumed.

	c. If the secondary resource is a non-stack, you
	   specify its type and a displayed name for it.

To have the above work, the generate_recipes_civs() proc was
rewritten to assign the secondary resource
to a recipe datum, which now has a list variable called
secondary_resource that will hold this information.

With the above system in place, some of the hard-coded
handling of checking for the existence of some secondary
resource, this secondary resource having an adequate amount
if it's a stack, and also using it up was removed.

This handling was then replaced with a more general way
of handling the checking for a secondary resource and
its usage that reads the needed information from
the recipe datum's secondary_resource list
if it exists.

Cloth and leather materials were given values into their
stack_type variables as they are needed
for the beforementioned system.

Procs that were checking the length of a recipe were
made to only check that the length isn't less than 13,
since a recipe can now have its length be greater than
13 when its secondary resource is described in the entry.

Some fixes not related to the recipe's
secondary resource were also performed.

Primitive wood armor was made craftable after adding the expected
RECIPE: part into its entries found in the recipe files.

The steppe_leather_helmet icon in hats.dmi was renamed
to leather_skullcap_helmet, because the helmet using this
icon state expected it to be named like this.

Two else if statements for the fire lance were combined
so that the spear checking logic is done after the
skill check.

Sandstone fortress walls should now be craftable with the
result type being set to /obj/covers/sandstone_wall/fortress
in the recipe files.

Sandstone arrowheads are now defined with some code added
in relation to their crafting and use.